refactor: delete unused dto, refactor service and controller to add Success Response and swagger annotations missing
