+++ src/main/java/com/femcoders/sitme/user/controller/UserController.java
@@ -4,12 +4,11 @@ package com.femcoders.sitme.user.controller;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.femcoders.sitme.shared.responses.SuccessResponse;
-import com.femcoders.sitme.space.dto.SpaceRequest;
-import com.femcoders.sitme.user.dtos.user.UserRequest;
 import com.femcoders.sitme.user.dtos.user.UserResponse;
-import com.femcoders.sitme.user.dtos.user.UserUpdateRequest;
-import com.femcoders.sitme.user.services.UserService;
+import com.femcoders.sitme.user.dtos.user.UserRequest;
+import com.femcoders.sitme.user.services.user.UserService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
@@ -22,7 +21,6 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController
@@ -33,30 +31,53 @@ public class UserController {
 
     private final UserService userService;
 
+    // GET ALL
+    @GetMapping
+    @PreAuthorize("hasRole('ADMIN')")
+    @Operation(
+            summary = "Get all users",
+            description = "Returns a list of all users. Only accessible to admins."
+    )
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "Users retrieved successfully"),
+            @ApiResponse(responseCode = "403", description = "Access denied")
+    })
+    @SecurityRequirement(name = "bearerAuth")
+    public ResponseEntity<SuccessResponse<List<UserResponse>>> getAllUsers() {
+
+        List<UserResponse> users = userService.getAllUsers();
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(SuccessResponse.of("Users list retrieved successfully", users));
+    }
+
+    // GET USER BY ID
     @GetMapping("/{id}")
-    public ResponseEntity<UserResponse> getUserById(
+    @PreAuthorize("hasRole('ADMIN')")
+    @Operation(
+            summary = "Get user by id",
+            description = "Returns a user profile by its id. Only accessible to admins."
+    )
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "User retrieved successfully"),
+            @ApiResponse(responseCode = "403", description = "Access denied")
+    })
+    @Parameter(description = "ID of the user profile", required = true)
+    @SecurityRequirement(name = "bearerAuth")
+    public ResponseEntity<SuccessResponse<UserResponse>> getUserById(
             @PathVariable Long id) {
+
         UserResponse userResponse = userService.getUserById(id);
-        return ResponseEntity.ok(userResponse);
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(SuccessResponse.of("User profile retrieved successfully", userResponse));
     }
 
-    @PostMapping("/{id}/image")
-    @Operation(summary = "Upload profile image for a user")
-    @ApiResponse(responseCode = "200", description = "Image uploaded successfully")
-    @ApiResponse(responseCode = "404", description = "User not found")
-    public ResponseEntity<UserResponse> uploadUserImage(
-            @PathVariable Long id,
-            @RequestParam("file") MultipartFile file
-    ) {
-
-        UserResponse userResponse = userService.uploadUserImage(id, file);
-        return ResponseEntity.ok(userResponse);
-    }
-
+    // UPDATE USER
     @PutMapping(value = "/{id}", consumes = "multipart/form-data")
     @PreAuthorize("hasRole('ADMIN')")
     @Operation(summary = "Update user",
-            description = "Allows an authenticated user to update their profile or admin to update any user")
+            description = "Allows an authenticated admin to update any user profile")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "User profile updated successfully"),
             @ApiResponse(responseCode = "400", description = "Invalid request data"),
@@ -70,7 +91,7 @@ public class UserController {
             @RequestPart(value = "file", required = false) MultipartFile file) throws JsonProcessingException {
 
         ObjectMapper mapper = new ObjectMapper();
-        UserUpdateRequest request = mapper.readValue(userJson, UserUpdateRequest.class);
+        UserRequest request = mapper.readValue(userJson, UserRequest.class);
 
         UserResponse updatedUser = userService.updateUser(id, request, file);
 
@@ -78,23 +99,24 @@ public class UserController {
                 .body(SuccessResponse.of("User profile updated successfully", updatedUser));
     }
 
-    @PutMapping("/profile/{id}")
-    @PreAuthorize("hasAnyRole('ADMIN','USER')")
-    @Operation(summary = "Update own profile",
-            description = "Allows an authenticated USER or ADMIN to update their own profile data")
+    // UPLOAD IMAGE
+    @PostMapping("/{id}/image")
+    @PreAuthorize("hasRole('ADMIN')")
+    @Operation(summary = "Upload profile image for a user")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Profile updated successfully"),
-            @ApiResponse(responseCode = "400", description = "Invalid request data"),
-            @ApiResponse(responseCode = "401", description = "Unauthorized"),
+            @ApiResponse(responseCode = "200", description = "Image uploaded successfully"),
             @ApiResponse(responseCode = "404", description = "User not found")
     })
     @SecurityRequirement(name = "bearerAuth")
-    public ResponseEntity<SuccessResponse<UserResponse>> updateProfile(
+    public ResponseEntity<SuccessResponse<UserResponse>> uploadUserImage(
             @PathVariable Long id,
-            @Valid @RequestBody UserUpdateRequest userUpdateRequest) {
+            @RequestParam("file") MultipartFile file
+    ) {
 
-        UserResponse updatedProfile = userService.updateProfile(id, userUpdateRequest);
-        return ResponseEntity.ok(SuccessResponse.of("Profile updated successfully", updatedProfile));
+        UserResponse userResponse = userService.uploadUserImage(id, file);
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(SuccessResponse.of("User profile updated successfully", userResponse));
     }
 
     @DeleteMapping("/{id}")
@@ -107,43 +129,26 @@ public class UserController {
     })
     @SecurityRequirement(name = "bearerAuth")
     public ResponseEntity<SuccessResponse<Void>> deleteUser(@PathVariable Long id) {
+
         userService.deleteUser(id);
+
         return ResponseEntity.status(HttpStatus.OK)
                 .body(SuccessResponse.of("User deleted successfully"));
     }
 
     @DeleteMapping("/{id}/image")
+    @PreAuthorize("hasRole('ADMIN')")
     @Operation(summary = "Delete user's profile image",
             description = "Deletes the profile image from Cloudinary and clears it in the database.")
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "Image deleted successfully"),
             @ApiResponse(responseCode = "404", description = "User not found")
     })
-    public ResponseEntity<SuccessResponse> deleteUserImage(@PathVariable Long id) {
+    public ResponseEntity<SuccessResponse<Void>> deleteUserImage(@PathVariable Long id) {
+
         userService.deleteUserImage(id);
 
-        return ResponseEntity.ok(
-                new SuccessResponse(true, "User image deleted successfully", LocalDateTime.now())
-        );
-    }
-
-    @GetMapping
-    @PreAuthorize("hasRole('ADMIN')")
-    @Operation(
-            summary = "Get all users",
-            description = "Resturns a list of all users. Only accessible to admins."
-    )
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "Users restrieved successfully"),
-            @ApiResponse(responseCode = "403", description = "Forbidden (insufficient role)"),
-            @ApiResponse(responseCode = "404", description = "No users found")
-    })
-    @SecurityRequirement(name = "bearerAuthh")
-    public ResponseEntity<List<UserResponse>> getAllUsers() {
-        List<UserResponse> users = userService.getAllUsers();
-        if (users.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        return ResponseEntity.ok(users);
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(SuccessResponse.of("User image deleted successfully"));
     }
 }
