+++ src/main/java/com/femcoders/sitme/user/dtos/user/UserRequest.java
@@ -1,16 +1,23 @@
 package com.femcoders.sitme.user.dtos.user;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 
 public record UserRequest(
         @NotBlank(message = "Name is required")
-        @Size(min = 2, max = 10, message = "Name must contain min 2 and max 10 characters")
+        @Size(min = 2, max = 50, message = "Name must contain min 2 and max 50 characters")
         String username,
+
         @NotBlank(message = "Email is required")
-        @Size(min = 10, max = 60, message = "Email must contain min 2 and max 60 characters")
+        @Email(message = "Invalid e-mail format",
+                regexp = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$")
         String email,
+
         @NotBlank(message = "Password is required")
+        @Pattern(message = "Password must contain a minimum of 8 characters, including a number, one uppercase letter, one lowercase letter and one special character",
+                regexp = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&+=.])(?=\\S+$).{8,}$")
         String password
 ) {
 }
