+++ src/main/java/com/femcoders/sitme/user/services/auth/UserAuthServiceImpl.java
@@ -1,4 +1,4 @@
-package com.femcoders.sitme.user.services;
+package com.femcoders.sitme.user.services.auth;
 
 import com.femcoders.sitme.email.EmailService;
 import com.femcoders.sitme.security.jwt.JwtService;
@@ -13,8 +13,6 @@ import com.femcoders.sitme.user.exceptions.IdentifierAlreadyExistsException;
 import com.femcoders.sitme.user.exceptions.InvalidCredentialsException;
 import com.femcoders.sitme.user.exceptions.UserNameNotFoundException;
 import com.femcoders.sitme.user.repository.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -22,7 +20,6 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
